# api/openapi.yaml
openapi: 3.1.0
info:
  title: BamForm API
  version: 1.0.0
  summary: Preventive maintenance record and approval system
  description: |
    Machine-readable contract for BamForm. This document is authoritative;
    BAMFORM-API-001 explains it. Divergence between this specification and the
    running implementation fails the CI build.

    Conventions, error catalogue, permission matrix and the offline sync
    protocol are described in BAMFORM-API-001.
  contact:
    name: BamForm project
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://form.bevorasg.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Admin]
      operationId: listUsers
      summary: List users
      description: ADMIN only (API_SPECIFICATION.md §10.9, UR-072).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: roleCode
          in: query
          schema: { $ref: '#/components/schemas/RoleCode' }
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Admin]
      operationId: createUser
      summary: Create a user
      description: |
        ADMIN only. `fullName`/`email`/`employeeId` are encrypted + blind-
        indexed (PR-106/107/108) — same real crypto path as `/auth/login`.
        `password` is admin-set (this slice's documented minimal mechanism —
        no invite/temp-password email flow exists yet). Rejects a duplicate
        `email` (blind-index uniqueness) with 409.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409':
          description: Duplicate email
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Admin]
      operationId: getUser
      summary: Get a user
      description: ADMIN only.
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Admin]
      operationId: updateUser
      summary: Update a user. Deactivation only, never deletion.
      description: |
        ADMIN only (UR-075). `active: false` is the only removal mechanism.
        `roleCodes`, when present, REPLACES the role set (produces a
        `permission_change` audit event, distinct from `update`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409':
          description: Duplicate email
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Ed25519-signed access token, 15 minute TTL
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      description: Opaque. Clients must not construct or parse it.
      schema: { type: string }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              pointer: { type: string }
              code: { type: string }
              message: { type: string }
      examples:
        - type: https://form.bevorasg.com/errors/incomplete-record
          title: Record is incomplete
          status: 422
          detail: 3 mandatory items have no recorded status.
          requestId: 01JQ8X2N
          errors:
            - pointer: /items/9
              code: REQUIRED
              message: Item 9 has no recorded status
    Page:
      type: object
      required: [page]
      properties:
        page:
          type: object
          required: [hasMore, limit]
          properties:
            nextCursor: { type: [string, 'null'] }
            hasMore: { type: boolean }
            limit: { type: integer }
            total: { type: [integer, 'null'], description: Present only when includeTotal=true }
...
    RoleCode:
      type: string
      description: The closed set `role.code` holds — seeded by migration, not created through the API (UR-073).
      enum: [MAINTAINER, TEAM_LEADER, ENGINEER, DOC_CONTROLLER, ADMIN, AUDITOR]
...
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, DEACTIVATED]
    User:
      type: object
      required: [id, fullName, email, status, active, roles, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        employeeId: { type: [string, 'null'] }
        fullName: { type: string }
        email: { type: string, format: email }
        status: { $ref: '#/components/schemas/UserStatus' }
        active: { type: boolean }
        roles:
          type: array
          items: { $ref: '#/components/schemas/RoleCode' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserCreate:
      type: object
      required: [fullName, email, password, roleCodes]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, format: email }
        employeeId: { type: string, minLength: 1, maxLength: 50 }
        password:
          type: string
          minLength: 12
          description: Admin-set initial password (this slice's documented minimal mechanism). Never returned by any endpoint.
        roleCodes:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/RoleCode' }
    UserUpdate:
      type: object
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, format: email }
        active: { type: boolean }
        roleCodes:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/RoleCode' }
...